src/main/java/com/edu/chmnu/ki_123/c3/ContactBook.java
package com.edu.chmnu.ki_123.c3;

import java.util.ArrayList;
import java.util.List;

public class ContactBook {
    private final List<Contact> contacts;

    public ContactBook() {
        contacts = new ArrayList<>();
    }

    public void addContact(Contact contact) {
        contacts.add(contact);
    }

    public void removeContact(Contact contact) {
        contacts.remove(contact);
    }

    public Contact searchContactByName(String name) {
        for (Contact contact : contacts) {
            if (contact.getName().equals(name)) {
                return contact;
            }
        }
        return null;
    }

    public List<Contact> getContacts() {
        return contacts;
    }
}